Business Case: Aerofit - Descriptive Statistics & Probability
There are no missing values in the data. There are 3 unique products in the dataset
KP281 is the most frequent product.  Minimum \& Maximum age of the person is 18 \& 50, mean is 28.79 and 75\% of persons have age less than or equal to 33}
Out of 180 data points, 104 are Male and the rest are female. 7- Standard deviation for Income \& Miles is very high. 
These variables might have outliers in it